fix #122950
