+++ src/vs/workbench/contrib/terminal/browser/terminal.ts
@@ -298,7 +298,7 @@ export interface ITerminalInstance {
 	onDisposed: Event<ITerminalInstance>;
 
 	onFocused: Event<ITerminalInstance>;
-	onDropped: Event<{ instance: ITerminalInstance, dragEvent: DragEvent; }>;
+	onDropped: Event<DragEvent>;
 	onProcessIdReady: Event<ITerminalInstance>;
 	onLinksReady: Event<ITerminalInstance>;
 	onRequestExtHostProcess: Event<ITerminalInstance>;
