+++ src/vs/workbench/contrib/terminal/browser/terminalService.ts
@@ -10,7 +10,6 @@ import { IDisposable } from 'vs/base/common/lifecycle';
 import { basename } from 'vs/base/common/path';
 import { isMacintosh, isWeb, isWindows, OperatingSystem } from 'vs/base/common/platform';
 import { URI } from 'vs/base/common/uri';
-import { DataTransfers } from 'vs/base/browser/dnd';
 import { FindReplaceState } from 'vs/editor/contrib/find/findState';
 import * as nls from 'vs/nls';
 import { ConfigurationTarget, IConfigurationService } from 'vs/platform/configuration/common/configuration';
@@ -661,35 +660,6 @@ export class TerminalService implements ITerminalService {
 		}));
 		instance.addDisposable(instance.onMaximumDimensionsChanged(() => this._onInstanceMaximumDimensionsChanged.fire(instance)));
 		instance.addDisposable(instance.onFocus(this._onActiveInstanceChanged.fire, this._onActiveInstanceChanged));
-		instance.addDisposable(instance.onDropped(async (e) => {
-			const dragEvent = e.dragEvent;
-			if (!dragEvent.dataTransfer) {
-				return;
-			}
-
-			// Check if files were dragged from the tree explorer
-			let path: string | undefined;
-			const resources = dragEvent.dataTransfer.getData(DataTransfers.RESOURCES);
-			if (resources) {
-				path = URI.parse(JSON.parse(resources)[0]).fsPath;
-			} else if (dragEvent.dataTransfer.files?.[0].path /* Electron only */) {
-				// Check if the file was dragged from the filesystem
-				path = URI.file(dragEvent.dataTransfer.files[0].path).fsPath;
-			}
-
-			if (!path) {
-				return;
-			}
-
-			const instance = e.instance;
-			if (!instance.shellLaunchConfig.executable) {
-				return;
-			}
-			const preparedPath = await this.preparePathForTerminalAsync(path, instance.shellLaunchConfig.executable, instance.title, instance.shellType, instance.isRemote);
-
-			instance.sendText(preparedPath, false);
-			instance.focus();
-		}));
 	}
 
 	registerProcessSupport(isSupported: boolean): void {
